+++ app/javascript/channels/index.js
@@ -1,3 +1,3 @@
 // Import all the channels to be used by Action Cable
 import "./chat_channel"
-import "./group_chat"
+import "./group_chat_channel"
